refactor: move all contexts to src/contexts
