+++ src/components/LineSeries.tsx
@@ -1,8 +1,8 @@
 import { line } from 'd3-shape';
 import { CSSProperties, useEffect, useMemo, useState } from 'react';
 
-import { useLegend } from '../legendContext';
-import { usePlotContext } from '../plotContext';
+import { useLegend } from '../contexts/legendContext';
+import { usePlotContext } from '../contexts/plotContext';
 import type { SeriesPoint } from '../types';
 import { getNextId, validateAxis } from '../utils';
